requirements.txt

fastapi
uvicorn==0.34.3
python-dotenv==1.1.0
motor==3.7.1
groq==0.27.0
pydantic[email]
fastapi==0.110.0
passlib